feat: remove NaiveStrategy
